[wpush] wpush 작업 진행중

+++ moka-web/moka-web-push/src/main/java/jmnet/moka/web/push/mvc/sender/repository/PushAppRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
  * 작업 Repository 2021. 2. 18.
  */
 @Repository
-public interface PushAppRepository extends JpaRepository<PushApp, Integer>, PushAppRepositorySupport {
+public interface PushAppRepository extends JpaRepository<PushApp, Long>, PushAppRepositorySupport {
 
 }
